Update test_cards.py
Overall, I don't believe that there are much changes needed, it seems pretty efficient! However, some ways you might be able to make it a little more efficient is by making the ordered deck and shuffled deck all be on the same line so it says ordered_deck, shuffled_deck = Deck() it might be useful.

Also the else that makes it reshuffle when a match is not made might also be useful, and worth checking out if you want. But overall, great job bro!